Cleanup analysis module & add plotting for eval.pkl
